Write map data to internal storage instead of external
On a real phone, it's not necessary to use external storage. It is even
bad, because it puts some files that are not relevant for the user on
the SD card, and requires an additional permission to do so.

Additionaly, the SD card emulation doesn't work on newer Android
emulators (version >=26).